+++ hooks/pre-commit-hook.py
@@ -1,40 +1,16 @@
 #!/usr/bin/python3
 import os
-import shutil
-import subprocess
-import sys
-from subprocess import Popen, PIPE
-
-def do_shell_action(commands):
-  ret = subprocess.Popen(commands, stdout=PIPE)
-  listOfBytes = ret.stdout.readlines()
-  ret = []
-  for line in listOfBytes:
-    l = line.decode("utf-8").replace('\n', '')
-    ret.append(l)
-  return ret
+import reformat
 
 #without this i was not allowed to do git diff --cached??
-do_shell_action(["git", "config", "--global", "--add", "safe.directory", "/workdir"])
+reformat.do_shell_action(["git", "config", "--global", "--add", "safe.directory", "/workdir"])
 
-changed_files=do_shell_action(["git", "diff", "--name-only", "--cached"])
-work_dir=do_shell_action(["git", "rev-parse", "--show-toplevel"])[0]
-
-def exists_clang_format_in_dir(dir):
-  for file in os.listdir(dir):
-    if file.endswith(".clang-format"):
-      return True
-  return False
-
-def find_clang_format_for_file(file):
-  file = os.path.join(work_dir, file)
-  directory_name = os.path.dirname(file)
-  while not exists_clang_format_in_dir(directory_name):
-    directory_name = os.path.dirname(directory_name)
-  return os.path.join(directory_name, ".clang-format")
+changed_files=reformat.do_shell_action(["git", "diff", "--name-only", "--cached"])
+work_dir=reformat.do_shell_action(["git", "rev-parse", "--show-toplevel"])[0]
 
 for file in changed_files:
   if file.endswith(".cpp") or file.endswith(".c") or file.endswith(".h") or file.endswith(".hpp"):
-    clang_format_file = find_clang_format_for_file(file)
-    do_shell_action(["clang-format-14", "-i", f"-style=file:{clang_format_file}", file])
-    do_shell_action(["git", "add", file])
+    file = os.path.join(work_dir, file)
+    clang_format_file = reformat.find_clang_format_for_file(file)
+    reformat.do_shell_action(["clang-format-14", "-i", f"-style=file:{clang_format_file}", file])
+    reformat.do_shell_action(["git", "add", file])
